pyproject.toml
[tool.poetry]
name = "todokist-app"
version = "0.1.0"
description = ""
authors = ["Павел <king-he@mail.ru>"]
readme = "README.md"
packages = [{include = "todokist_app"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.4.0"
django = "^4.1.5"
django-environ = "^0.9.0"
psycopg2 = "^2.9.5"
gunicorn = "^20.1.0"
drf-spectacular = "^0.25.1"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"
dataclasses-json = "^0.5.7"
pydantic = "^1.10.4"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
